requirements.txt
dj-database-url==0.4.1
Django==1.9.8
django-filter==0.13.0
djangorestframework==3.4.1
psycopg2==2.6.2
whitenoise==3.2
